refactor: products-array-dialog to shared module
