+++ src/app/shared/dialogs/products-array/products-array-dialog.component.css
